# docs/key-files/backend-api-spec.yaml
openapi: 3.0.3
info:
  title: MebelPlace API
  description: |
    MebelPlace Backend API - платформа для торговли мебелью с видео-контентом, 
    заявками, чатами и системой подписок.
    
    ## Основные возможности:
    - Регистрация и аутентификация пользователей
    - Загрузка и просмотр видео с мебелью
    - HLS (HTTP Live Streaming) для адаптивного видео
    - Multi-variant HLS с разными качествами (360p, 720p, 1080p)
    - Автоматическое создание thumbnail и HLS сегментов
    - Асинхронная обработка видео с отслеживанием статуса
    - WebRTC интеграция для live streaming
    - Система заявок и предложений
    - Чаты и групповые чаты
    - Каналы и подписки
    - Система уведомлений
    - Звонки и голосовые комнаты
    - AR/3D модели
    - Аналитика и геймификация
    
    ## Rate Limiting:
    - HLS endpoints: 100 запросов/минуту на пользователя
    - Live streaming: 10 запросов/минуту на пользователя
    - Thumbnail: 200 запросов/минуту на пользователя
    - Общие API: 1000 запросов/минуту на пользователя
    
  version: 2.4.0
  contact:
    name: MebelPlace Support
    email: support@mebelplace.com.kz
  license:
    name: Proprietary
servers:
  - url: https://mebelplace.com.kz/api/v2
    description: Production server
  - url: http://localhost:8080/api/v2
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /integrations/payments/providers:
    get:
      tags: [Integrations]
      summary: Получить подключенных провайдеров
      description: Получить список подключенных платежных провайдеров
      responses:
        '200':
          description: Список провайдеров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentProvider'
        '401':
          description: Неавторизован

  # Реферальная система
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://mebelplace.com.kz/auth/oauth/authorize
          tokenUrl: https://mebelplace.com.kz/auth/oauth/token
          scopes:
            read: Чтение данных
            write: Запись данных
            admin: Административные права
  schemas:
...
    PaymentProvider:
      type: object
      properties:
        id:
          type: integer
        provider:
          type: string
        is_connected:
          type: boolean
        connected_at:
          type: string
          format: date-time
        settings:
          type: object
...